Main pages for the administrator and the employee
Admin Page:
Once the admins are logged in they view a list of the app's users.
They can click on a user in order to edit their info or create a
new one from scratch.

User Page:
Here, users view a list of all the applications they've submitted so far.
They can also make a new request for a leave of absence.

	new file:   src/welcome.php